backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.1
manim==0.19.0
numpy==1.26.4
Pillow==10.2.0 